+++ NEWS.md
@@ -1,3 +1,4 @@
+* Added `phi_constants` table to output of `get_abs_paths()`.
 * Exemplar table now sits in Mapping folder.
   Changed `get_abs_paths()` and tests to reflect this change.